requirements.txt
flask
mysql-connector-python
gunicorn
python-dotenv